+++ tests/Junty/Runner/RunnerTest.php
@@ -8,56 +8,57 @@
  
 namespace Test\Junty;
 
-use Junty\{TaskInterface, Runner, AbstractTask};
+use Junty\Runner\Runner;
+use Junty\{TaskInterface, AbstractTask};
 use Junty\Stream\StreamHandler;
 
 class RunnerTest extends \PHPUnit_Framework_TestCase
 {
     public function testCreatingAndExecutingATask()
     {
-        $tasker = new Runner();
+        $runner = new Runner();
 
-        $tasker->task('task_1', function () {
+        $runner->task('task_1', function () {
             $_SERVER['FOO'] = 'bar';
         });
 
-        $tasker->run();
+        $runner->run();
 
         $this->assertEquals($_SERVER['FOO'], 'bar');
     }
 
     public function testOrderingTasks()
     {
-        $tasker = new Runner();
+        $runner = new Runner();
 
-        $tasker->task('task_1', function () {
+        $runner->task('task_1', function () {
             $_SERVER['BARZ'] = 'bar';
         });
 
-        $tasker->task('task_2', function () {
+        $runner->task('task_2', function () {
             $_SERVER['BARZ'] = 'show';
         });
 
-        $tasker->task('task_3', function () {
+        $runner->task('task_3', function () {
             $_SERVER['BARZ'] = 'kbz';
         });
 
-        $tasker->order('task_3', 'task_1', 'task_2');
+        $runner->order('task_3', 'task_1', 'task_2');
 
-        $tasker->run();
+        $runner->run();
 
         $this->assertEquals($_SERVER['BARZ'], 'show');
     }
 
     public function testReturnOfTaskMethodIsTheTask()
     {
-        $tasker = new Runner();
+        $runner = new Runner();
         $name = 'task_1';
         $callback = function () {
             $_SERVER['FOO'] = 'bar';
         };
 
-        $return = $tasker->task($name, $callback);
+        $return = $runner->task($name, $callback);
 
         $this->assertEquals($return->getName(), $name);
         $this->assertEquals($return->getCallback(), $callback);
@@ -65,9 +66,9 @@ class RunnerTest extends \PHPUnit_Framework_TestCase
 
     public function testCreatingTaskByInstanceOfATask()
     {
-        $tasker = new Runner();
+        $runner = new Runner();
 
-        $tasker->task(new class() extends AbstractTask {
+        $runner->task(new class() extends AbstractTask {
             public function getName() : string
             {
                 return 'task_1';
@@ -81,9 +82,9 @@ class RunnerTest extends \PHPUnit_Framework_TestCase
             }
         });
 
-        $tasker->order('task_1');
+        $runner->order('task_1');
 
-        $tasker->run();
+        $runner->run();
 
         $this->assertEquals($_SERVER['CHUBBY'], 'bunny');
     }
@@ -92,16 +93,43 @@ class RunnerTest extends \PHPUnit_Framework_TestCase
     {
         $_this = null;
 
-        $tasker = new Runner();
+        $runner = new Runner();
 
-        $tasker->task('task_1', function () use (&$_this) {
+        $runner->task('task_1', function () use (&$_this) {
             $_this = $this;
         });
 
-        $tasker->order('task_1');
+        $runner->order('task_1');
 
-        $tasker->run();
+        $runner->run();
 
         $this->assertTrue($_this instanceof StreamHandler);
+    }
+
+    public function testIfGetterForTasksReturnsAllTasks()
+    {
+        $runner = new Runner();
+
+        $task2CallbackReturn = 'hiiii';
+
+        $runner->task('task_1', function () {});
+        $runner->task('task_2', function () use ($task2CallbackReturn) {
+            return $task2CallbackReturn;
+        });
+
+        $tasks = $runner->getTasks();
+
+        $this->assertCount(2, $tasks);
+
+        $this->assertArrayHasKey('task_1', $tasks);
+        $this->assertArrayHasKey('task_2', $tasks);
+        
+        foreach ($tasks as $task) {
+            $this->assertTrue($task instanceof TaskInterface);
+        }
+
+        $cb2 = $tasks['task_2']->getCallback();
+
+        $this->assertEquals($cb2(), $task2CallbackReturn);
     }
 }