+++ CHANGELOG.md
@@ -2,7 +2,10 @@ Changelog
 =========
 
 ## HEAD
---
+* Update on runner:
+  * Created the namespace ```Junty\Runner```.
+  * Added method ```getTasks``` to ```Junty\Runner\RunnerInterface```.
+  * Updated console logging to show what task is beeing executed.
 
 ## v0.1.0
-* Released
+* Released.